Vi treng LeesahOpplysningstype også der vi handterer PDL-hendingane, ikkje berre der vi siler dei, så flyttar til pdl-modellen vår så vi kan gjenbruke enum heller enn å basere oss på magiske tekststrengar
