+++ apps/etterlatte-hendelser-pdl/src/test/kotlin/hendelserpdl/MeldingSendtPaaRapid.kt
@@ -1,6 +1,7 @@
 package no.nav.etterlatte.hendelserpdl
 
 import com.fasterxml.jackson.annotation.JsonProperty
+import no.nav.etterlatte.pdl.hendelse.LeesahOpplysningstype
 
 data class MeldingSendtPaaRapid<T>(
     @JsonProperty("@event_name") val eventName: String,
