+++ libs/etterlatte-pdl-model/src/main/kotlin/pdlhendelse/LeesahOpplysningstype.kt
@@ -1,0 +1,11 @@
+package no.nav.etterlatte.pdl.hendelse
+
+enum class LeesahOpplysningstype {
+    ADRESSEBESKYTTELSE_V1,
+    FORELDERBARNRELASJON_V1,
+    UTFLYTTING_FRA_NORGE,
+    DOEDSFALL_V1,
+    VERGEMAAL_ELLER_FREMTIDSFULLMAKT_V1,
+    SIVILSTAND_V1,
+    BOSTEDSADRESSE_V1,
+}
